requirements.txt
openai==0.27.2
pinecone_client==2.2.1
streamlit==1.18.0
